gcp_functions/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-aiplatform
cryptography
